Split up UriViewer module
