+++ src/components/data/FileBrowser.js
@@ -11,7 +11,7 @@ import { NameModal } from 'src/components/NameModal'
 import { UploadProgressModal } from 'src/components/ProgressBar'
 import RequesterPaysModal from 'src/components/RequesterPaysModal'
 import { FlexTable, HeaderCell, TextCell } from 'src/components/table'
-import UriViewer from 'src/components/UriViewer'
+import { UriViewer } from 'src/components/UriViewer/UriViewer'
 import { Ajax } from 'src/libs/ajax'
 import colors from 'src/libs/colors'
 import { reportError, withErrorReporting } from 'src/libs/error'
